RequestResponse Unit Test
